Move tree.py to utils
Why:

The tree node will be used for PBT and probably for trial objects as
well in the future so it must be generalized.
